Добавил пагинацию и фильтры

+++ Models/EntryModel.cs
@@ -7,7 +7,7 @@ namespace DynamicSun_weather.Models
         public int ID { get; set; }
         public DateTime Date { get; set; }
         public float? Temp { get; set; }
-        public int? Humidity { get; set; }
+        public float? Humidity { get; set; }
         public float? Dew { get; set; }
         public int? Pressure { get; set; }
         public string? WindDirection { get; set; }
